Fixes for re-enabling more MSVC level 4 warnings: misc edition #3
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Possibly-uninitialized local variable

These are the warnings that have appeared in the codebase since I began
uploading my first cleanup pass a few weeks ago.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/400313002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284378 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e3ee14ea097fc271907013b09ef61685c8a0c0b6

+++ cc/surfaces/surface_aggregator_unittest.cc
@@ -31,7 +31,7 @@ namespace {
 
 SurfaceId InvalidSurfaceId() {
   static SurfaceId invalid;
-  invalid.id = -1;
+  invalid.id = static_cast<uint64_t>(-1);
   return invalid;
 }
